+++ lib/runtime/server-side-rendering/helpers/test-logger.ts
@@ -4,6 +4,7 @@ export const testLogger: Logger = {
     // tslint:disable:no-console
     debug: (msg) => console.log(msg),
     info: (msg) => console.log(msg),
+    warn: (msg) => console.log(msg),
     error: (...args: any[]) => {
         console.log(args)
     },