+++ lib/runtime/server.ts
@@ -7,5 +7,6 @@
 
 export * from './server/assets'
 export * from './server/server'
+export * from './universal'
 
 export { default as config } from './config/config'