+++ lib/runtime/util/log.ts
@@ -1,6 +1,7 @@
 export interface Logger {
     debug(msg: string | { msg: string, [key: string]: any }): void
     info(msg: string | { msg: string, [key: string]: any }): void
+    warn(msg: string | { msg: string, [key: string]: any }): void
     error(msg: string): void
     error<T extends { err: Error, msg: string }>(obj: T): void
 }